Consolidate NativeMethods.MSG

+++ src/System.Windows.Forms/src/System/Windows/Forms/ComboBox.cs
@@ -4042,7 +4042,7 @@ namespace System.Windows.Forms
                     break;
 
                 default:
-                    if (m.Msg == NativeMethods.WM_MOUSEENTER)
+                    if (m.Msg == (int)NativeMethods.WM_MOUSEENTER)
                     {
                         DefWndProc(ref m);
                         OnMouseEnterInternal(EventArgs.Empty);
